Ps 133 change post url redirection

+++ gateway/src/main/java/com/patreonshout/jpa/CreatorPageRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
  * Spring Data Repository for easy use of CRUD operations on the {@link com.patreonshout.beans.CreatorPage} object
  */
 @Repository
-public interface CreatorPageRepository extends JpaRepository<CreatorPage, Long> {
+public interface CreatorPageRepository extends JpaRepository<CreatorPage, Integer> {
 
 }
